+++ src/clusters/BaseClusterWorker.ts
@@ -7,11 +7,12 @@ interface Setup {
 }
 
 export class BaseClusterWorker {
-    bot: any;
-    clusterID: number;
-    workerID: number;
-    ipc: IPC;
-    shutdown?: Function;
+    public bot: any;
+    public clusterID: number;
+    public workerID: number;
+    public ipc: IPC;
+    /** Function called for graceful shutdown of the cluster */
+    public shutdown?: Function;
 
     public constructor(setup: Setup) {
         this.bot = setup.bot;