+++ src/clusters/Cluster.ts
@@ -17,6 +17,7 @@ export class Cluster {
     private token!: string;
     app!: BaseClusterWorker;
     shutdown?: Boolean;
+    fetchTimeout!: number;
 
     constructor() {
         //@ts-ignore
@@ -55,6 +56,7 @@ export class Cluster {
                         this.clientOptions = message.clientOptions;
                         this.token = message.token;
                         this.whatToLog = message.whatToLog;
+                        this.fetchTimeout = message.fetchTimeout;
 
                         if (this.shards < 0) return;
                         this.connect();
@@ -243,5 +245,8 @@ export class Cluster {
         //let App = (await import(this.path)).default;
         //App = App.default ? App.default : App;
         this.app = new App({bot: this.bot, clusterID: this.clusterID, workerID: worker.id});
+        
+        // Add fetch timeout to ipc
+        this.app.ipc.fetchTimeout = this.fetchTimeout;
     }
 }